reflect "connected" status in a named semaphore for keys requesting this
- new `connection-semname` key parameter in `authorized_keys` gives the
name ("ABC") of a semaphore to create.  This will maintain semaphore
"/dev/shm/sem.ABC" as long as the key is "connected"; i.e. as long as
there is a process running on the server which was launched as a result
of authentication with the given key.  When used with `ClientKeepAlive`,
this seems to me the natural definition; in any case it is the useful one
for the sensorgnome project.

- deals correctly with `-oControlMaster` and independent logins using
the same key, regardless of the number of distinct clients.

Note: requires linking sshd with `-lrt -lpthread`

Closes #2.